+++ tunafish/components/connection.js
@@ -1,25 +1,14 @@
 import axios from 'axios';
-import { useEffect, useState } from 'react';
 import { useToast } from '@leafygreen-ui/toast';
 import TextInput from "@leafygreen-ui/text-input";
 import Button from "@leafygreen-ui/button";
 import { PasswordInput } from "@leafygreen-ui/password-input";
 
-function MongoDBConnection({connection,handleConnectionChange,connected,setConnected}) {
-    const { pushToast, popToast, clearStack } = useToast();
-    // const [error, setError] = useState(null);
-    // const [toasts, setToasts] = useState({warning:'',progress:'',success:''})
-    // // const [errorToast, setErrorToast] = useState('');
-    // // const [progressToast, setProgressToast] = useState('');
-    // // const [successToast, setSuccessToast] = useState('false');
+function MongoDBConnection({connection,handleConnectionChange,setIndexes}) {
+    const { pushToast, clearStack } = useToast();
 
     const uri = connection?.uri? connection.uri:undefined;
     const namespace = (connection?.database && connection?.collection)? `${connection.database}.${connection.collection}`:undefined;
-
-    useEffect(()=>{
-        clearStack();
-    },[connection]);
-    
     
     const handleSubmit = () => {
         pushToast({variant:"progress",title:"Connecting",description:`Trying to connect to ${connection.database}.${connection.collection}`});
@@ -27,17 +16,25 @@ function MongoDBConnection({connection,handleConnectionChange,connected,setConne
             .then(resp => {
                 clearStack();
                 pushToast({variant:"success",title:"Connected!",description:`Successfully connected to ${connection.database}.${connection.collection}`}); 
-                setConnected(true);
+                // setConnected(true);
+                pushToast({variant:"progress",title:"Fetching indexes",description:`Retrieving search indexes from ${connection.database}.${connection.collection}`});
+                fetchIndexes(connection).then(resp=>{
+                    setIndexes(resp.data);
+                    pushToast({variant:"success",title:"Fetched indexes",description:`Got ${resp.data.length} search indexes from ${connection.database}.${connection.collection}`}); 
+                })
+                .catch(error=>{
+                    clearStack();
+                    pushToast({timeout:0,variant:"warning",title:"Search failure",description:`Failed to get indexes from ${connection.database}.${connection.collection}. ${error}`})
+                });
             })
             .catch(error => {
                 clearStack();
+                // setConnected(false);
                 pushToast({timeout:0,variant:"warning",title:"Failed",description:`Connection to ${connection.database}.${connection.collection} failed. ${error}`})
             })
     }
 
     return (
-
-   
         <div style={{
             display: "grid",
             gridTemplateColumns: "40% 25% 90px",
@@ -63,6 +60,13 @@ function MongoDBConnection({connection,handleConnectionChange,connected,setConne
 function connect(conn) {
     return new Promise((resolve,reject) => {
         axios.post(`api/post/atlas-search/index/connect?`,{connection:conn})
+        .then(response => resolve(response))
+        .catch((error) => reject(error.response.data))
+    });
+}
+function fetchIndexes(conn) {
+    return new Promise((resolve,reject) => {
+        axios.post(`api/post/atlas-search/index/list?`,{connection:conn})
         .then(response => resolve(response))
         .catch((error) => reject(error.response.data))
     });
